Refactor pages scripts into another file
