+++ extension/src/content-script/osu/pages/match.ts
@@ -1,7 +1,7 @@
 // https://osu.ppy.sh/community/matches/110067650
 
 import { TFlagItems, addFlagUser, addFlagUsers } from "@src/content-script/osu/flagHtml";
-import { idFromProfileUrl, nextAbortControllerSignal } from "./content";
+import { idFromProfileUrl, nextAbortControllerSignal } from "../content";
 
 export const updateFlagsMatches = async () => {
     if (!location.href.includes("osu.ppy.sh/community/matches/")) return;
